refactor: 컴포넌트 페이지 폴더 정리
