+++ src/domain/order/components/pay-contents/DeliveryInfo.tsx
