+++ src/pages/order-confirm/OrderConfirmPage.tsx
@@ -1,8 +1,8 @@
 import { useNavigate } from "react-router-dom";
 import { Header } from "../../components/common";
 import BackArrowButton from "../../components/common/assets/BackArrowButton";
-import OrderContents from "../../domain/order/components/order-contents/OrderContents";
-import PayContents from "../../domain/order/components/order-contents/pay-contents/PayContents";
+import OrderContents from "./components/OrderContents";
+import PayContents from "./components/PayContents";
 import { CouponProvider } from "./context/CouponProvider";
 import { useScrollToTop } from "../../hooks/useScrollToTop";
